+++ packages/cli/src/helpers.ts
@@ -7,6 +7,7 @@ import {
 } from '@convergence-rfq/sdk';
 import { Command } from 'commander';
 
+import { Connection } from '@solana/web3.js';
 import { Instrument } from './types';
 import { DEFAULT_KEYPAIR_FILE, DEFAULT_RPC_ENDPOINT } from './constants';
 import { Opts } from './cvg';
@@ -97,4 +98,14 @@ export const extractBooleanString = (opts: Opts, name: string): boolean => {
   }
 
   return value === 'true' ? true : false;
+};
+
+export const getSigConfirmation = async (
+  connection: Connection,
+  tx: string
+) => {
+  const result = await connection.getSignatureStatus(tx, {
+    searchTransactionHistory: true,
+  });
+  return result?.value?.confirmationStatus;
 };