+++ packages/js/CHANGELOG.md
@@ -12,6 +12,12 @@
 
 - Hxro integration
 
+## 4.5.30
+
+### Patch Changes
+
+- add UpdateBaseAsset operation to sdk
+
 ## 4.5.29
 
 ### Patch Changes